플레이어 움직이게 애니메이션 제작

+++ BlockCodingProject/Assets/01. Scripts/JoyStick/JoyStickInput.cs
@@ -11,8 +11,8 @@ public class JoyStickInput : MonoBehaviour, IPointerDownHandler, IPointerEnterHa
     [HideInInspector]
     public bool joyStickHold = false;
 
-    public static float horizontalRaw = 0;
-    public static float verticalRaw = 0;
+    public static int horizontalRaw = 0;
+    public static int verticalRaw = 0;
 
     public Sprite[] joystickes;
